Initial ldap_auth module
This commit is the initial functioning framework for the auth_ldap module. This will be
expanded out to adequately cover more options and systems. This commit is admittedly rough,
lacking things like rspec-puppet tests and documentation.

Reasoning: Why this module? How does it help?

The goal of this module is to allow any host using the name service switch have the ability
to lookup passwd databases and others from LDAP. This is typically done with nss_ldap or
pam_ldap. Sometimes both, however distributions of Linux are starting to move over to using
nslcd for lookups due to its flexibility and extended features. There are a lot of similar
features and directives between these systems. This module should make it painless to manage
ldap based user auth in environments where different families of *nix or different versions
of the same distro are being used.